changed approach for fetching data from Mongo shell
Previous approach was bad because after sending command to Mongo shell, application waited 100 milliseconds and it read the result (even if query still hasn't terminated or result wasn't competely sent to STDOUT). That caused occasional crashes or incomplete list of data. This was very common when working with remote database or when the result-set was pretty big.
In new approach is generated pseudo-random string that marks the end of result (it has role of delimiting identifier in here-document/here-string), so result is being waited and read (periodically, e.g. every 100 milliseconds) until previously mentioned generated-string is read.