it is now possible to simultaneously add multiple objects with auto-id
Until now, it was possible to "simultaneously" (to have opened multiple tabs in the same time) add/edit multiple documents for whom user manually specified id (value of "_id" property). Now it is also possible to open multiple tab-windows for document-insertion even if for some (or all of them) is selected that value of "_id" property will be auto-generated by the MongoDB server.